Adicionado mapa e criado menu para utilizador após o login

+++ app/src/main/java/com/example/praticalcity/api/Situacao.kt
@@ -2,8 +2,8 @@ package com.example.praticalcity.api
 
 data class Situacao(
     val id: Int,
-    val latitude: Double,
-    val longitude: Double,
+    val latitude: Float,
+    val longitude: Float,
     val titulo: String,
     val descr: String,
     val foto: String
